requirements.txt
discord.py==2.3.1
openai==0.27.8
python-dotenv==1.0.0
PyYAML==6.0.1
